requirements.txt
customtkinter
SpeechRecognition
pyttsx3
deep-translator
langdetect
pyautogui
google-generativeai
gtts
pygame
